Improve scripts
