+++ package.json
@@ -12,29 +12,34 @@
     "sushiswap"
   ],
   "scripts": {
-    "ci": "ts-node ./src/utils/initServer/create-indexes.ts",
-    "it": "ts-node ./src/utils/initServer/init-tokens.ts",
-    "ip": "ts-node ./src/utils/initServer/init-pools.ts",
-    "iw": "ts-node ./src/utils/initServer/init-whales.ts",
-    "lcs": "ts-node-dev ./src/listener-commons.ts",
-    "lm": "ts-node-dev ./src/listener-mempool.ts",
-    "lc": "ts-node-dev ./src/listener-confirmation.ts",
-    "web": "ts-node-dev ./src/_websocket-server.ts",
-    "build": "cd client && yarn build && cd .. && tsc",
-    "set-heroku": "ts-node ./src/utils/initServer/set-heroku.ts",
-    "dev-d-txs": "ts-node ./src/utils/dev-utils/drop-transactions.ts",
-    "dev-r-txs": "ts-node ./src/utils/dev-utils/reset-transactions.ts",
-    "dev-r-txc": "ts-node ./src/utils/dev-utils/reset-transactions-c.ts",
-    "dev-r-txp`": "ts-node ./src/utils/dev-utils/reset-transactions-p.ts",
-    "dev-r-t-h": "ts-node ./src/utils/dev-utils/reset-transactions-t-h.ts",
-    "dev-r-pools": "ts-node ./src/utils/dev-utils/reset-pools.ts",
-    "dev-r-blocks": "ts-node ./src/utils/dev-utils/reset-blocks.ts",
-    "dev-r-tokens": "ts-node ./src/utils/dev-utils/reset-tokens.ts",
-    "pretty": "prettier --config .prettierrc 'src/**/*.ts' --write",
-    "web-worker": "node ./dist/_websocket-server.js",
-    "worker-a": " node ./dist/listener-commons.js",
-    "worker-b": "node ./dist/listener-mempool.js",
-    "worker-c": "node ./dist/listener-confirmation.js"
+    "create-indexes": "ts-node ./src/utils/initServer/create-indexes.ts",
+    "init-tokens": "ts-node ./src/utils/initServer/init-tokens.ts",
+    "init-pools": "ts-node ./src/utils/initServer/init-pools.ts",
+    "init-whales": "ts-node ./src/utils/initServer/init-whales.ts",
+    
+    "dev-drop-txs": "ts-node ./src/utils/dev-utils/drop-transactions.ts",
+    "dev-reset-txs": "ts-node ./src/utils/dev-utils/reset-transactions.ts",
+    "dev-reset-pools": "ts-node ./src/utils/dev-utils/reset-pools.ts",
+    "dev-reset-blocks": "ts-node ./src/utils/dev-utils/reset-blocks.ts",
+    "dev-reset-tokens": "ts-node ./src/utils/dev-utils/reset-tokens.ts",
+   
+    "dev-websocket": "ts-node-dev ./src/websocket-server.ts",
+    "dev-mempool": "ts-node-dev ./src/listener-mempool.ts",
+    "dev-blocks": "ts-node-dev ./src/listener-blocks.ts",
+    "dev-confirmed": "ts-node-dev ./src/listener-confirmed.ts",
+    "dev-fix-db": "ts-node ./src/fix-db.ts",
+    
+    "build-all": "yarn build-client && yarn build-backend",
+    "build-client": "cd client && yarn build",
+    "build-backend": "tsc",
+
+    "websocket": "node ./dist/websocket-server.js",
+    "mempool": " node ./dist/listener-mempool.js",
+    "blocks": "node ./dist/listener-blocks.js",
+    "confirmed": "node ./dist/listener-confirmed.js",
+    "fix-db": "node ./dist/fix-db.js",
+
+    "pretty": "prettier --config .prettierrc 'src/**/*.ts' --write"
   },
   "devDependencies": {
     "@types/express": "^4.17.12",
